// src/types/api.ts
export interface Character {
  id: number;
  name: string;
  status: 'Alive' | 'Dead' | 'unknown';
  species: string;
  type: string;
  gender: 'Female' | 'Male' | 'Genderless' | 'unknown';
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface CharacterResponse {
  info: {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null;
  };
  results: Character[];
}

export interface FilterParams {
  status?: string;
  gender?: string;
  page?: number;
}

export type StatusFilter = 'alive' | 'dead' | 'unknown' | '';
export type GenderFilter = 'female' | 'male' | 'genderless' | 'unknown' | '';